+++ plone/app/upgrade/tests/base.py
@@ -9,13 +9,13 @@ from plone.app.testing.bbb import PTC_FIXTURE
 from plone.app.testing.bbb import PloneTestCase
 from Products.CMFCore.interfaces import IActionCategory
 from Products.CMFCore.interfaces import IActionInfo
-from Products.CMFCore.tests.base.testcase import WarningInterceptor
 from Products.CMFCore.utils import getToolByName
 from Products.GenericSetup.context import TarballImportContext
 from zope.configuration import xmlconfig
 from zope.site.hooks import setSite
 
 import transaction
+import warnings
 
 #
 # Base TestCase for upgrades
@@ -150,7 +150,7 @@ class MigrationTest(PloneTestCase):
             skins.addSkinSelection(skin, ','.join(path))
 
 
-class FunctionalUpgradeTestCase(PloneTestCase, WarningInterceptor):
+class FunctionalUpgradeTestCase(PloneTestCase):
 
     _setup_fixture = 0
     site_id = 'test'
@@ -168,9 +168,8 @@ class FunctionalUpgradeTestCase(PloneTestCase, WarningInterceptor):
 
     def importFile(self, context, name):
         path = join(abspath(dirname(context)), 'data', name)
-        self._trap_warning_output()
-        self.app._importObjectFromFile(path, verify=0)
-        self._free_warning_output()
+        with warnings.catch_warnings():
+            self.app._importObjectFromFile(path, verify=0)
 
     def migrate(self):
         oldsite = getattr(self.app, self.site_id)